Fixed bug: changed int in get, set to double

+++ src/main/java/com/example/bgmmixer/model/Stage.java
@@ -46,7 +46,7 @@ public class Stage {
         return startTime;
     }
 
-    public void setStartTime(int startTime) {
+    public void setStartTime(double startTime) {
         this.startTime = startTime;
     }
 
@@ -54,7 +54,7 @@ public class Stage {
         return endTime;
     }
 
-    public void setEndTime(int endTime) {
+    public void setEndTime(double endTime) {
         this.endTime = endTime;
     }
